#11 acabat (implementació + testos)

+++ src/main/java/edu/upc/essi/gps/ecommerce/DiscountService.java
@@ -80,7 +80,7 @@ public class DiscountService {
         return discountRepository.findById(id);
     }
 
-    public List<Discount> listByTriggerId(int id) {
+    public List<Discount> listByTriggerId(long id) {
         return discountRepository.findByTriggerId(id);
     }
